requirements.txt
dice==3.*
discord==1.*
discord.py==1.3.4
python-dotenv==0.14.0